:recycle: move tests
